cadd-vault-backend-scripts/requirements.txt
# requirements.txt - Updated with test dependencies

# Core dependencies (existing)
supabase>=1.0.0
python-dotenv>=0.19.0
httpx>=0.24.0
habanero>=1.2.0
paperscraper>=0.2.0
backoff>=2.2.0
rich>=12.0.0

# New dependencies for improved functionality
tenacity>=8.0.0
pydantic>=1.10.0
pandas>=1.5.0
openpyxl>=3.1.0

# Test dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-html>=3.1.0
pytest-cov>=4.0.0

# Optional: For more detailed test reporting
pytest-xdist>=3.0.0  # Run tests in parallel
pytest-mock>=3.10.0  # Additional mocking capabilities